main/serializers/user_profile.py
from rest_framework import serializers

from main.models.user_profile import UserProfile


class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = [
            "id",
            "image",
        ]
